api/dev-requirements.txt
-i https://pypi.org/simple
astroid==2.2.5
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
django-rest-swagger==2.2.0
djangorestframework==3.9.3
factory-boy==2.11.1
faker==1.0.5
idna==2.8
isort==4.3.17
itypes==1.1.0
jinja2==2.10.1
lazy-object-proxy==1.3.1
markupsafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
pep8==1.7.1
pycodestyle==2.5.0
pylint==2.3.1
python-dateutil==2.8.0
requests==2.21.0
simplejson==3.16.0
six==1.12.0
text-unidecode==1.2
typed-ast==1.3.5 ; implementation_name == 'cpython'
uritemplate==3.0.0
urllib3==1.24.2
wrapt==1.11.1
